requirements.txt
playwright==1.58.0
openai==2.32.0
pydantic==2.13.3
PyYAML==6.0.2
Jinja2==3.1.4
loguru==0.7.3
pytest==8.3.4
pytest-mock==3.14.0
